my order screen finish

+++ lib/feature/order/data/model/response/get_order.dart
@@ -116,7 +116,7 @@ class OrdersCartItems {
 @JsonSerializable()
 class OrdersProduct {
   String? title;
-  int? ratingsAverage;
+  double? ratingsAverage;
   String? sId;
   String? image;
